APP-626: [Extension] Cannot scroll history in detail token
