+++ src/components/Header/Header.enhance.tsx
@@ -14,7 +14,6 @@ export interface IProps {
     selectAccount?: boolean;
     canSearch?: boolean;
     customHeader?: React.FunctionComponent | React.ReactElement;
-    removeMarginTop?: boolean;
 }
 
 export interface IMergeProps extends TInner, IProps {}
